+++ keyring/examples/keyring.rs
@@ -3,7 +3,8 @@ use keyring::Entry;
 const KEY: &str = "key";
 const VAL: &str = "val";
 
-fn main() {
+#[tokio::main]
+async fn main() {
     env_logger::builder().is_test(true).init();
 
     // set global keyring service name
@@ -11,14 +12,14 @@ fn main() {
 
     // set entry secret
     let entry = Entry::from(KEY);
-    entry.set_secret(VAL).unwrap();
+    entry.set_secret(VAL).await.unwrap();
 
     // get entry secret
-    let val = entry.get_secret().unwrap();
+    let val = entry.get_secret().await.unwrap();
     assert_eq!(entry.get_key(), KEY);
     assert_eq!(val, VAL);
 
     // delete entry
-    entry.delete_secret().unwrap();
-    assert_eq!(entry.find_secret().unwrap(), None);
+    entry.delete_secret().await.unwrap();
+    assert_eq!(entry.find_secret().await.unwrap(), None);
 }