+++ keyring/CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0] - 2023-12-11
+
+### Changed
+
+- Made `Entry` serializable and deserializable using `serde`.
+- `Entry::{get,find,set,delete}_secret` are now `async`.
+
 ## [0.2.0] - 2023-12-10
 
 ### Added